+++ R/coerce.R
@@ -22,6 +22,9 @@ setMethod(
   f = "as_composition",
   signature = c(from = "matrix"),
   definition = function(from) {
+    ## Validation
+    arkhe::assert_type(from, "numeric")
+
     ## Make row/column names
     lab <- make_names(x = NULL, n = nrow(from), prefix = "S")
     rownames(from) <- if (has_rownames(from)) rownames(from) else lab
@@ -31,8 +34,7 @@ setMethod(
     totals <- rowSums(from, na.rm = TRUE)
     from <- from / totals
 
-    grp <- as_groups(rep(NA, nrow(from)))
-    .CompositionMatrix(from, totals = unname(totals), groups = grp)
+    .CompositionMatrix(from, totals = unname(totals))
   }
 )
 
@@ -42,27 +44,25 @@ setMethod(
 setMethod(
   f = "as_composition",
   signature = c(from = "data.frame"),
-  definition = function(from, parts = NULL, groups = NULL,
+  definition = function(from, parts = NULL, groups = NULL, autodetect = TRUE,
                         verbose = getOption("nexus.verbose")) {
     ## Clean row/column names
     lab <- make_names(x = NULL, n = nrow(from), prefix = "S")
     rownames(from) <- if (has_rownames(from)) rownames(from) else lab
     colnames(from) <- make_names(x = colnames(from), n = ncol(from), prefix = "V")
 
-    ## Group names
-    grp <- rep(NA_character_, nrow(from))
-    if (!is.null(groups)) grp <- from[, groups, drop = FALSE]
-    grp <- as_groups(grp)
-
     ## Remove non-numeric columns
     if (is.null(parts)) {
-      parts <- arkhe::detect(from, f = is.double, margin = 2) # Logical
-      if (verbose) {
-        n <- sum(parts)
-        what <- ngettext(n, "part", "parts")
-        cols <- paste0(colnames(from)[parts], collapse = ", ")
-        msg <- "Found %g %s (%s)."
-        message(sprintf(msg, n, what, cols))
+      if (isTRUE(autodetect)) {
+        parts <- arkhe::detect(from, f = is.numeric, margin = 2)
+        if (isTRUE(verbose)) {
+          n <- sum(parts)
+          what <- ngettext(n, "Found %g part (%s)", "Found %g parts (%s)")
+          cols <- paste0(colnames(from)[parts], collapse = ", ")
+          message(sprintf(what, n, cols))
+        }
+      } else {
+        arkhe::assert_filled(parts)
       }
     } else {
       if (is.numeric(parts)) parts <- seq_len(ncol(from)) %in% parts
@@ -76,7 +76,12 @@ setMethod(
     totals <- rowSums(coda, na.rm = TRUE)
     coda <- coda / totals
 
-    .CompositionMatrix(coda, totals = unname(totals), groups = grp)
+    z <- .CompositionMatrix(coda, totals = unname(totals))
+    if (is.null(groups)) return(z)
+
+    ## Group names
+    grp <- from[, groups, drop = FALSE]
+    group(z, by = grp, verbose = verbose)
   }
 )
 
@@ -93,48 +98,76 @@ setMethod(
 )
 
 # To data.frame ================================================================
-# @export
-# @rdname augment
-# @aliases augment,CompositionMatrix-method
-# setMethod(
-#   f = "augment",
-#   signature = c(x = "CompositionMatrix"),
-#   definition = function(x) {
-#     data.frame(
-#       .group = groups(x),
-#       x
-#     )
-#   }
-# )
-
-# @export
-# @rdname augment
-# @aliases augment,LogRatio-method
-# setMethod(
-#   f = "augment",
-#   signature = c(x = "LogRatio"),
-#   definition = function(x) {
-#     data.frame(
-#       .group = groups(x),
-#       x
-#     )
-#   }
-# )
-
 #' @method as.data.frame CompositionMatrix
 #' @export
-as.data.frame.CompositionMatrix <- function(x, ...) {
-  as.data.frame(methods::as(x, "matrix"), row.names = rownames(x))
+as.data.frame.CompositionMatrix <- function(x, row.names = rownames(x),
+                                            optional = FALSE, ...) {
+  as.data.frame(methods::as(x, "matrix"), row.names = row.names, optional = optional)
 }
+
+#' @export
+#' @rdname as.data.frame
+#' @aliases as.data.frame,CompositionMatrix-method
+setMethod("as.data.frame", "CompositionMatrix", as.data.frame.CompositionMatrix)
+
+#' @method as.data.frame GroupedComposition
+#' @export
+as.data.frame.GroupedComposition <- function(x, row.names = rownames(x),
+                                             optional = FALSE, ...,
+                                             group_var = ".group") {
+  z <- data.frame(
+    methods::as(x, "matrix"),
+    row.names = row.names,
+    check.names = !optional
+  )
+  z[[group_var]] <- group_names(x)
+  z
+}
+
+#' @export
+#' @rdname as.data.frame
+#' @aliases as.data.frame,GroupedComposition-method
+setMethod("as.data.frame", "GroupedComposition", as.data.frame.GroupedComposition)
 
 #' @method as.data.frame LogRatio
 #' @export
-as.data.frame.LogRatio <- function(x, ...) {
-  as.data.frame(methods::as(x, "matrix"), row.names = rownames(x))
+as.data.frame.LogRatio <- function(x, row.names = rownames(x),
+                                   optional = FALSE, ...) {
+  as.data.frame(methods::as(x, "matrix"), row.names = row.names, optional = optional)
 }
+
+#' @export
+#' @rdname as.data.frame
+#' @aliases as.data.frame,LogRatio-method
+setMethod("as.data.frame", "LogRatio", as.data.frame.LogRatio)
+
+#' @method as.data.frame GroupedLogRatio
+#' @export
+as.data.frame.GroupedLogRatio <- function(x, row.names = rownames(x),
+                                          optional = FALSE, ...,
+                                          group_var = ".group") {
+  z <- data.frame(
+    methods::as(x, "matrix"),
+    row.names = row.names,
+    check.names = !optional
+  )
+  z[[group_var]] <- group_names(x)
+  z
+}
+
+#' @export
+#' @rdname as.data.frame
+#' @aliases as.data.frame,GroupedLogRatio-method
+setMethod("as.data.frame", "GroupedLogRatio", as.data.frame.GroupedLogRatio)
 
 #' @method as.data.frame OutlierIndex
 #' @export
-as.data.frame.OutlierIndex <- function(x, ...) {
-  as.data.frame(x@standard, row.names = rownames(x))
+as.data.frame.OutlierIndex <- function(x, row.names = rownames(x),
+                                       optional = FALSE, ...) {
+  as.data.frame(x@standard, row.names = row.names, optional = optional)
 }
+
+#' @export
+#' @rdname as.data.frame
+#' @aliases as.data.frame,OutlierIndex-method
+setMethod("as.data.frame", "OutlierIndex", as.data.frame.OutlierIndex)